Refactor: Git is initialized once, in the root of app
