+++ src/git.ts
@@ -1,0 +1,24 @@
+import simpleGit, { SimpleGit, SimpleGitOptions } from "simple-git";
+import * as vscode from "vscode";
+
+export function initGit(): SimpleGit | undefined {
+    // TODO: Support multiple repositories in the same VSCode Workspace?
+    let gitRepoPath: string = "";
+    if (
+        vscode.workspace.workspaceFolders &&
+        vscode.workspace.workspaceFolders.length > 0
+    ) {
+        gitRepoPath = vscode.workspace.workspaceFolders[0].uri.fsPath;
+    }
+    if (!gitRepoPath) {
+        return
+    }
+
+    console.log("init sturdy", gitRepoPath);
+    const options: SimpleGitOptions = {
+        baseDir: gitRepoPath,
+        binary: "git",
+        maxConcurrentProcesses: 6,
+    };
+    return simpleGit(options);
+}
